кусок расширенного поиска (без категорий книг)

+++ db/migrate/20151211180548_create_pages.rb
@@ -3,7 +3,7 @@ class CreatePages < ActiveRecord::Migration
     create_table :pages do |t|
 
       t.references  :book,            index:true, foreign_key:true, null:false
-      t.integer     :internal_order,  null:false
+      t.decimal     :internal_order,  null:false
       t.string      :url_name,        null:false
       t.text        :page_text,       null:false
